🍀🍀 Remove __pycache__ and refactoring show messages
